requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
pyyaml>=5.4.0
pillow>=9.0.0
tqdm>=4.62.0
psutil>=5.8.0
scipy>=1.9.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pandas>=1.3.0
pytest>=7.0.0

torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.30.0
accelerate>=0.20.0
xformers>=0.0.22

peft>=0.10.0
lpips>=0.1.4
clean-fid>=0.1.35
einops>=0.6.1
open-clip-torch>=2.20.0
timm>=0.9.2
dominate>=2.9.0

evo>=1.23.0
